Completed implementation of handlebars, restructured folder structure (namely the views folder) so our files could be seen. However, there seems to be an error when reading html/hbs files
I finished setting up handlebars in the package.json file (it was missing one line)

 I also did some re-structuring of the setup in order to correctly our pages be read (it seems that there needs to be a "views" folder? [It kept trying to search for one so I just created it and moved what we had in there].

Now it can actually find our files html/hbs (it is correctly redirecting to the page I set it to [which is loginStudent, rn for testing purposes but we can change that later] but I ran into another problem?
I keep getting this error in the photo. It seems to be an encoding error as I have no idea what the two mysterious characters are? (I attached my code so you can check as well). On the brightside, it seems to be at least finding and reading the right file (or attempting to), which I struggled with for a bit.

So in short just note that my commit seems to have fixed some problems but it is still flawed.